refactor: input fields
